+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -181,7 +181,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-02T15:12:47.774Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-03T06:58:46.742Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")