+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6JQQJ0Q3GJN0J175BXVZA2W"
+        "01K6MDVPQ7V0HAB3XYGW4G3PGX"
     )
     r"""The id of the resource"""