+++ packages/orq-rc/RELEASES.md
@@ -3109,3 +3109,13 @@ Based on:
 - [python v3.14.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.5] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.5 - packages/orq-rc
+
+## 2025-10-03 07:00:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.1 (2.721.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.6] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.6 - packages/orq-rc